+++ packages/main/src/Form.ts
@@ -47,6 +47,7 @@ type GroupItemsInfo = {
 	groupItem: IFormItem,
 	classes: string,
 	items: Array<ItemsInfo>,
+	accessibleNameRef: string | undefined
 }
 
 type ItemsInfo = {
@@ -162,6 +163,13 @@ type ItemsInfo = {
  *   | 2 | 5 |
  *   | 3 | 6 |
  *   ```
+ *
+ * ### Keyboard Handling
+ *
+ * - [Tab] - Moves the focus to the next interactive element within the Form/FormGroup (if available) or to the next element in the tab chain outside the Form
+ * - [Shift] + [Tab] - Moves the focus to the previous interactive element within the Form/FormGroup (if available) or to the previous element in the tab chain outside the Form
+ * - [F6] - Moves the focus to the first interactive element of the next FormGroup (if available) or to the next element in the tab chain outside the Form
+ * - [Shift] + [F6] - Moves the focus to the first interactive element of the previous FormGroup (if available) or to the previous element in the tab chain outside the Form
  *
  * ### ES6 Module Import
  *
@@ -301,6 +309,8 @@ class Form extends UI5Element {
 	onAfterRendering() {
 		// Create additional CSS for number of columns that are not supported by default.
 		this.createAdditionalCSSStyleSheet();
+
+		this.setFastNavGroup();
 	}
 
 	setColumnLayout() {
@@ -335,6 +345,14 @@ class Form extends UI5Element {
 			item.labelSpan = this.labelSpan;
 			item.itemSpacing = this.itemSpacing;
 		});
+	}
+
+	setFastNavGroup() {
+		if (this.hasGroupItems) {
+			this.removeAttribute("data-sap-ui-fastnavgroup");
+		} else {
+			this.setAttribute("data-sap-ui-fastnavgroup", "true");
+		}
 	}
 
 	setGroupsColSpan() {
@@ -398,8 +416,12 @@ class Form extends UI5Element {
 		return !!this.header.length;
 	}
 
-	get ariaLabelledByID(): string | undefined {
+	get effectiveАccessibleNameRef(): string | undefined {
 		return this.hasCustomHeader ? undefined : `${this._id}-header-text`;
+	}
+
+	get effectiveAccessibleRole(): string | undefined {
+		return this.hasGroupItems ? "region" : "form";
 	}
 
 	get groupItemsInfo(): Array<GroupItemsInfo> {
@@ -428,6 +450,7 @@ class Form extends UI5Element {
 
 			return {
 				groupItem,
+				accessibleNameRef: (groupItem as FormGroup).headerText ? `${groupItem._id}-group-header-text` : undefined,
 				classes: `ui5-form-column-spanL-${groupItem.colsL} ui5-form-column-spanXL-${groupItem.colsXl} ui5-form-column-spanM-${groupItem.colsM} ui5-form-column-spanS-${groupItem.colsS}`,
 				items: this.getItemsInfo((Array.from(groupItem.children) as Array<IFormItem>)),
 			};